Refactored the app by clustering components together and the personal model stuff together.
