+++ mesa/src/personal_model/activity_level.js
